+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Added
 - Add field setting to restrict which column types editors can choose ([#53](https://github.com/verbb/tablemaker/issues/53)).
+- Add field settings for min/max rows and min/max columns ([#38](https://github.com/verbb/tablemaker/issues/38)).
 
 ### Removed
 - Removed Column Label field setting.